Fix: Move DbOperation types from database.h to _plugins.h
The DbItemType, DbOperationType, and DbOperation struct were defined in
database.h, an internal header not distributed with the plugin SDK. Plugins
receiving CB_DBOPERATION callbacks could not access the struct fields without
it.

Move the full definitions into _plugins.h so plugins can use them directly.
database.h now includes _plugins.h, keeping all existing internal code
unaffected.
